Add `destroy` for cleaning up global handle.
Since the CUBLAS handle should be destroyed before CUDArt resets the
device (which happens in CUDArt's `close`), it should be possible for a
library user to explicitly close the handle.

Providing a function for this avoids the library user forgetting to set
it to `C_NULL` and thus creating double-destroys `atexit`.